Add files via upload
Login + logout + zabezpieczenia

+++ Articles/article.php
@@ -3,6 +3,8 @@
 require 'includes/db_connection.php';
 require 'includes/article-functions.php';
 
+session_start();
+
 $conn = getDBconn();
 
 if (isset($_GET['id'])) {
@@ -27,16 +29,18 @@ if (isset($_GET['id'])) {
               <p><?= htmlspecialchars($article['content']); ?></p>
             </article>
             <br>
-            <a href="edit-article.php?id=<?= $article['id']; ?>">Edit article</a>
-              <!-- PRÓBOWAŁEM UMIEŚCIĆ BUTTON, ALE NIE MOGĘ ZMUSIĆ GO ŻEBY POBRAŁ ID -->
-            <!-- <form action="edit-article.php?id=<?= $article['id']; ?>" method="get">
-              <button>Edit article</button>
-            </form> -->
-            <br><br>
-            <!-- <form  action="delete-article.php?id=<?= $article['id']; ?>" method="post">
-              <button>Delete article</button>
-            </form> -->
-            <!-- TO DZIAŁA ALE CHCIAŁEM, ŻEBY BYŁO SPÓJNIE -->
-            <a href="delete-article.php?id=<?= $article['id']; ?>">Delete article</a>
+            <?php if (isset($_SESSION['is_logged_in']) && $_SESSION['is_logged_in']): ?>
+              <a href="edit-article.php?id=<?= $article['id']; ?>">Edit article</a>
+                <!-- PRÓBOWAŁEM UMIEŚCIĆ BUTTON, ALE NIE MOGĘ ZMUSIĆ GO ŻEBY POBRAŁ ID -->
+              <!-- <form action="edit-article.php?id=<?= $article['id']; ?>" method="get">
+                <button>Edit article</button>
+              </form> -->
+              <br><br>
+              <!-- <form  action="delete-article.php?id=<?= $article['id']; ?>" method="post">
+                <button>Delete article</button>
+              </form> -->
+              <!-- TO DZIAŁA ALE CHCIAŁEM, ŻEBY BYŁO SPÓJNIE -->
+              <a href="delete-article.php?id=<?= $article['id']; ?>">Delete article</a>
+            <?php endif; ?>
     <?php endif; ?>
  <?php require 'includes/footer.php' ?>
